Flush stream and sync filesystem
In testing we found that powering down the drone can cause the json
files to be zero bytes when the drone is powered back on again.  This
causes the drone to not know what radio settings to set.  The flush and
sync force the file to be written to disk and powering down the drone
and powering it back on again the file is always there on disk.